Fix a few crashes

+++ linuxdoom-1.10/r_draw.c
@@ -26,12 +26,12 @@
 static const char rcsid[] = "$Id: r_draw.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 
 #include "doomdef.h"
+#include "r_local.h"
 
 #include "i_system.h"
 #include "w_wad.h"
 #include "z_zone.h"
-
-#include "r_local.h"
+#include <stdint.h>
 
 // Needs access to LFB (guess what).
 #include "v_video.h"
@@ -426,7 +426,7 @@ void R_InitTranslationTables(void)
 	int i;
 
 	translationtables = Z_Malloc(256 * 3 + 255, PU_STATIC, 0);
-	translationtables = (byte*)(((int)translationtables + 255) & ~255);
+	translationtables = (byte*)(((intptr_t)translationtables + 255) & ~255);
 
 	// translate just the 16 green colors
 	for (i = 0; i < 256; i++)
